issue #192 refactor: utilize getTimeString function in ScheduleItem
